+++ packages/react-components/react-avatar/src/stories/AvatarGroup/index.stories.beta.tsx
@@ -4,6 +4,7 @@ import descriptionMd from './AvatarGroupDescription.md';
 import bestPracticesMd from './AvatarGroupBestPractices.md';
 
 export { Default } from './AvatarGroupDefault.stories';
+export { Layout } from './AvatarGroupLayout.stories';
 
 export default {
   title: 'Preview Components/AvatarGroup',