+++ packages/react-components/react-avatar/src/stories/AvatarGroupItem/AvatarGroupItemDefault.stories.tsx
@@ -1,4 +1,4 @@
 import * as React from 'react';
-import { AvatarGroupItem, AvatarGroupItemProps } from '../../index';
+import { AvatarGroupItem, AvatarGroupItemProps } from '../../AvatarGroupItem';
 
 export const Default = (props: Partial<AvatarGroupItemProps>) => <AvatarGroupItem name="Katri Athokas" {...props} />;