+++ packages/react-components/react-card/src/components/Card/useCardStyles.ts
@@ -5,17 +5,23 @@ import { cardHeaderClassNames } from '../CardHeader/useCardHeaderStyles';
 import { cardFooterClassNames } from '../CardFooter/useCardFooterStyles';
 import type { CardSlots, CardState } from './Card.types';
 import type { SlotClassNames } from '@fluentui/react-utilities';
+import { createFocusOutlineStyle } from '@fluentui/react-tabster';
 
+/**
+ * Static CSS class names used internally for the component slots.
+ */
 export const cardClassNames: SlotClassNames<CardSlots> = {
   root: 'fui-Card',
 };
-export const cardCSSVars = {
-  cardSizeVar: '--fui-Card--size',
-};
 
 /**
- * Styles for the root slot
+ * CSS variable names used internally for uniform styling in Card.
  */
+export const cardCSSVars = {
+  cardSizeVar: '--fui-Card--size',
+  cardBorderRadiusVar: '--fui-Card--border-radius',
+};
+
 const useStyles = makeStyles({
   root: {
     display: 'flex',
@@ -23,7 +29,7 @@ const useStyles = makeStyles({
     ...shorthands.overflow('hidden'),
     color: tokens.colorNeutralForeground1,
 
-    // Border setting using after pseudo element to allow CardPreview to render behind it
+    // Border setting using after pseudo element to allow CardPreview to render behind it.
     '::after': {
       position: 'absolute',
       top: 0,
@@ -35,8 +41,10 @@ const useStyles = makeStyles({
 
       ...shorthands.borderStyle('solid'),
       ...shorthands.borderWidth(tokens.strokeWidthThin),
+      ...shorthands.borderRadius(`var(${cardCSSVars.cardBorderRadiusVar})`),
     },
 
+    ...shorthands.borderRadius(`var(${cardCSSVars.cardBorderRadiusVar})`),
     ...shorthands.padding(`var(${cardCSSVars.cardSizeVar})`),
     ...shorthands.gap(`var(${cardCSSVars.cardSizeVar})`),
 
@@ -48,6 +56,14 @@ const useStyles = makeStyles({
     [`> :not(.${cardPreviewClassNames.root}):not(.${cardHeaderClassNames.root}):not(.${cardFooterClassNames.root})`]: {
       flexGrow: 1,
     },
+
+    ...createFocusOutlineStyle({
+      style: {
+        outlineRadius: `var(${cardCSSVars.cardBorderRadiusVar})`,
+        outlineWidth: tokens.strokeWidthThick,
+      },
+      selector: 'focus',
+    }),
   },
 
   orientationHorizontal: {
@@ -84,15 +100,15 @@ const useStyles = makeStyles({
 
   sizeSmall: {
     [cardCSSVars.cardSizeVar]: '8px',
-    ...shorthands.borderRadius(tokens.borderRadiusSmall),
+    [cardCSSVars.cardBorderRadiusVar]: tokens.borderRadiusSmall,
   },
   sizeMedium: {
     [cardCSSVars.cardSizeVar]: '12px',
-    ...shorthands.borderRadius(tokens.borderRadiusMedium),
+    [cardCSSVars.cardBorderRadiusVar]: tokens.borderRadiusMedium,
   },
   sizeLarge: {
     [cardCSSVars.cardSizeVar]: '16px',
-    ...shorthands.borderRadius(tokens.borderRadiusLarge),
+    [cardCSSVars.cardBorderRadiusVar]: tokens.borderRadiusLarge,
   },
 
   interactiveNoOutline: {
@@ -213,7 +229,7 @@ const useStyles = makeStyles({
 });
 
 /**
- * Apply styling to the Card slots based on the state
+ * Apply styling to the Card slots based on the state.
  */
 export const useCardStyles_unstable = (state: CardState): CardState => {
   const styles = useStyles();