+++ packages/react-components/theme-designer/src/components/Sidebar/Sidebar.tsx
@@ -6,7 +6,7 @@ import { AddCircleRegular } from '@fluentui/react-icons';
 export interface SidebarProps {
   className?: string;
   keyColor: string;
-  changeKeyColor: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  setKeyColor: React.Dispatch<React.SetStateAction<string>>;
 }
 
 const useStyles = makeStyles({
@@ -76,6 +76,9 @@ export const Sidebar: React.FC<SidebarProps> = props => {
   const [darkTheme, setDarkTheme] = React.useState<string>('#000000');
   const changeDarkTheme = React.useCallback(e => setDarkTheme(e.target.value), [setDarkTheme]);
 
+  const handleOnBlur = React.useCallback(e => props.setKeyColor(e.target.value), [props]);
+  const handleOnChange = React.useCallback(e => props.setKeyColor(e.target.value), [props]);
+
   return (
     <div className={mergeClasses(styles.root, props.className)}>
       <TabList className={styles.tabs} size="medium" defaultSelectedValue="use">
@@ -96,16 +99,10 @@ export const Sidebar: React.FC<SidebarProps> = props => {
               appearance="underline"
               id={keyColorId}
               value={props.keyColor}
-              onChange={props.changeKeyColor}
+              onChange={handleOnChange}
             />
             <div className={styles.colorPicker} style={{ backgroundColor: props.keyColor }}>
-              <input
-                className={styles.color}
-                type="color"
-                id={keyColorId}
-                value={props.keyColor}
-                onChange={props.changeKeyColor}
-              />
+              <input className={styles.color} type="color" id={keyColorId} onBlur={handleOnBlur} />
             </div>
           </div>
         </div>
@@ -122,6 +119,7 @@ export const Sidebar: React.FC<SidebarProps> = props => {
             />
             <div className={styles.colorPicker} style={{ backgroundColor: lightTheme }}>
               <input
+                disabled={true}
                 className={styles.color}
                 type="color"
                 id={lightThemeId}
@@ -137,6 +135,7 @@ export const Sidebar: React.FC<SidebarProps> = props => {
             <Input size="small" appearance="underline" id={darkThemeId} value={darkTheme} onChange={changeDarkTheme} />
             <div className={styles.colorPicker} style={{ backgroundColor: darkTheme }}>
               <input
+                disabled={true}
                 className={styles.color}
                 type="color"
                 id={darkThemeId}